Fixes & universal article form component & mutation error handling - change admin user can not be deleted - admin privileges are not shown in table - error toast shown when invalid json response

+++ api/src/users/UsersGateway.php
@@ -79,6 +79,11 @@ class UsersGateway
 
     public function delete(string $id)
     {
+        $user = $this->get($id);
+        if ($user["username"] == "admin") {
+            throw new Exception("Admin user can not be deleted!");
+        }
+
         $sql = "DELETE FROM users WHERE id = :id";
         $stmt = $this->conn->prepare($sql);
